fix: philosopher didn't wake neighbor in semaphore
When add cast to the Dinner class in semaphore solution I copied the
same code to all places that needed it, causing to refer only the
current philosopher, when in the return_cutlery function it must
refer to the neighbor that will be wake.
Fixed the Dinner object definition in the main class.